PR: Support re-using sensors in z_thermal_adjust
Module: z_thermal_adjust
Multiple `z_thermal_adjust` sections may now be defined in config to compensate for different sources of z thermal expansion that happen at different rates. Heater objects, which encapsulate a temperature sensor, may be used as a sensor source to drive the adjustment. This allows an Extruder to drive an adjustment to compensate for nozzle expansion. This is a common problem with nozzle probes that need to use a low temperature while probing.

There is an explanation and test data in this thread: https://klipper.discourse.group/t/z-thermal-adjust-for-nozzles/22457

Signed-off-by: Gareth Farrington <gareth@waves.ky>