Tap Recorder Module
Adds a `[tap_recorder]` TapClassifier implementation. This saves save tap data to a file. Recording can be turned on and off via gcode commands. It can also delegate tap classification to another classifier module, allowing you to record and validate the performance of that module. Data is stored in JSONL format for efficiency when collecting large datasets.

Signed-off-by: Gareth Farrington <gareth@waves.ky>