Use a RetrySession in ProbePointsHelper
If using the CIRCLE stratey and retrying multiple times at the same points (e.g. for QGL) any fouled points are rememberd and avoided by keeping a session between attempts (calls to `start_probe`). The number of retries are reset for all points on each pass but the fouled points are saved.


Signed-off-by: Gareth Farrington <gareth@waves.ky>