Expose a generate_points() API on BedMesh
This change allows callers to create tools that leverage the bed mesh config (e.g. for complex things like faulty regions) and GCode parameters without having to actually probe those points. The generated points can then be used for other purposes. E.g. to visit each point for a calibration routine.

Unfortunatly this changes the internal state of the bed mesh. The current code is not conducive to reuse. However on the next bed mesh call the state will be reset to whatever bed mesh the user specifies.

Signed-off-by: Gareth Farrington <gareth@waves.ky>